Move Seed words to wallet tab
