+++ src/containers/Settings/MiningSettings.tsx
@@ -1,13 +1,12 @@
 import CpuMiningMarkup from './sections/mining/CpuMiningMarkup.tsx';
 import GpuMiningMarkup from './sections/mining/GpuMiningMarkup.tsx';
-import SeedWordsMarkup from './sections/mining/SeedWordsMarkup';
+import SeedWordsMarkup from './sections/wallet/SeedWordsMarkup/index.ts';
 
 export const MiningSettings = () => {
     return (
         <>
             <CpuMiningMarkup />
             <GpuMiningMarkup />
-            <SeedWordsMarkup />
         </>
     );
 };
