+++ src/containers/Settings/WalletSettings.tsx
@@ -1,11 +1,13 @@
 import WalletAddressMarkup from './sections/wallet/WalletAddressMarkup.tsx';
 import MoneroAddressMarkup from './sections/wallet/MoneroAddressMarkup';
+import SeedWordsMarkup from './sections/wallet/SeedWordsMarkup/SeedWordsMarkup.tsx';
 
 export const WalletSettings = () => {
     return (
         <>
             <WalletAddressMarkup />
             <MoneroAddressMarkup />
+            <SeedWordsMarkup />
         </>
     );
 };
