+++ src/containers/Settings/sections/wallet/SeedWordsMarkup/SeedWords.tsx
