+++ src/containers/Settings/sections/wallet/SeedWordsMarkup/SeedWordsMarkup.tsx
