+++ src/containers/Settings/sections/wallet/SeedWordsMarkup/index.ts
