+++ src/containers/Settings/sections/wallet/SeedWordsMarkup/useGetSeedWords.ts
